requirements.txt
openai
faiss-cpu
numpy
streamlit
python-dotenv
jira